Move instances to a separate file. Add companion methods for functor and applicative.
