+++ kategory/src/main/kotlin/kategory/arrow/Function1.kt
@@ -18,38 +18,12 @@ data class Function1<in A, out R>(val f: (A) -> R) : FunctionInject<R>, Function
     class F private constructor()
 
     companion object {
-        fun <P> monad() = object : Function1MonadReader<P> {}
+        fun <P> functor() = object : Function1Instances<P> {}
 
-        fun <P> monadReader() = object : Function1MonadReader<P> {}
+        fun <P> applicative() = object : Function1Instances<P> {}
+
+        fun <P> monad() = object : Function1Instances<P> {}
+
+        fun <P> monadReader() = object : Function1Instances<P> {}
     }
-}
-
-interface Function1MonadReader<P> : MonadReader<Function1.F, P> {
-
-    override fun ask(): HK<Function1.F, P> =
-            { a: P -> a }.k()
-
-    override fun <A> local(f: (P) -> P, fa: HK<Function1.F, A>): Function1<P, A> =
-            f.andThen { fa.ev().invoke(it) }.k()
-
-    override fun <A> pure(a: A): Function1<P, A> =
-            { _: P -> a }.k()
-
-    override fun <A, B> map(fa: HK<Function1.F, A>, f: (A) -> B): HK<Function1.F, B> =
-            f.compose { b: B -> fa.ev().invoke(b) }.k()
-
-    override fun <A, B> flatMap(fa: HK<Function1.F, A>, f: (A) -> HK<Function1.F, B>): Function1<P, B> =
-            Function1 { p -> f(fa.ev().invoke(p)).ev().invoke(p) }
-
-    override fun <A, B> tailRecM(a: A, f: (A) -> HK<Function1.F, Either<A, B>>): Function1<P, B> =
-            Function1 { p ->
-                tailrec fun loop(thisA: A): B =
-                        f(thisA).ev().invoke(p).fold({ loop(it) }, { it })
-
-                loop(a)
-            }
-
-    fun monad(): Monad<Function1.F> = this
-
-    fun monadReader(): MonadReader<Function1.F, P> = this
 }
